requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
Django==5.2.9
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==38.2.0
pillow==12.0.0
PyJWT==2.10.1
RapidFuzz==3.14.3
redis==7.1.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
